refactor: refactored responses tab
